Completed all serverside interaction.
The server can now continuously accept new connections, and sends chat messages to the server using your IP address as a username. Simple server-client interaction is demonstrated in ChatServer.java, which may be modified along with ClientHandler.java in order to make the actual game (along with other new classes for the game instances themselves, and the entire clientside version).